# docs/swagger.yaml
definitions:
  common.JsonResult:
    properties:
      code:
        type: integer
      data:
        type: object
      msg:
        type: string
    type: object
  models.Cluster:
    properties:
      category:
        type: string
      clusterName:
        type: string
      remark:
        type: string
      status:
        type: integer
    type: object
  models.Node:
    properties:
      clusterId:
        type: integer
      clusterName:
        type: string
      ip:
        type: string
      nodeType:
        description: 1 主机主节点，2 集群主节点，3，集群从节点
        type: integer
      password:
        type: string
      port:
        type: integer
      remark:
        type: string
      status:
        description: 1正常，2非正常
        type: integer
      username:
        type: string
    type: object
  models.Project:
    properties:
      clusterName:
        type: string
      remark:
        type: string
    type: object
info:
  contact: {}
  license: {}
paths:
  /cluster/createCluster:
    post:
      consumes:
      - application/json
      description: 创建集群
      parameters:
      - description: Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Cluster'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.JsonResult'
  /cluster/deleteCluster:
    delete:
      consumes:
      - application/json
      description: 删除集群
      parameters:
      - description: Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Cluster'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.JsonResult'
  /cluster/listClusters:
    post:
      consumes:
      - application/json
      description: 查询所有集群
      parameters:
      - description: Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Cluster'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.JsonResult'
  /cluster/listNodes:
    post:
      consumes:
      - application/json
      description: 查寻集群下面的所有主机
      parameters:
      - description: Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Node'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.JsonResult'
  /cluster/pageClusters:
    post:
      consumes:
      - application/json
      description: 分页查询集群
      parameters:
      - description: Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Cluster'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.JsonResult'
  /cluster/removeNode:
    delete:
      consumes:
      - application/json
      description: 移除集群的主机节点
      parameters:
      - description: Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Node'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.JsonResult'
  /cluster/setNode:
    post:
      consumes:
      - application/json
      description: 为集群添加主机节点
      parameters:
      - description: Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Node'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.JsonResult'
  /cluster/updateCluster:
    put:
      consumes:
      - application/json
      description: 更新集群
      parameters:
      - description: Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Cluster'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.JsonResult'
  /node/createNode:
    post:
      consumes:
      - application/json
      description: 创建主机节点
      parameters:
      - description: Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Node'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.JsonResult'
  /node/deleteNode:
    delete:
      consumes:
      - application/json
      description: 删除主机节点
      parameters:
      - description: Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Node'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.JsonResult'
  /node/pageNodes:
    post:
      consumes:
      - application/json
      description: 删除主机节点
      parameters:
      - description: Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Node'
      - description: excludeInCluster 排除已经加入集群的节点
        in: query
        name: action
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.JsonResult'
  /node/updateNode:
    put:
      consumes:
      - application/json
      description: 更新主机节点
      parameters:
      - description: Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Node'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.JsonResult'
  /project/createProject:
    post:
      consumes:
      - application/json
      description: 创建项目
      parameters:
      - description: Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Project'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.JsonResult'
  /project/deleteProject:
    delete:
      consumes:
      - application/json
      description: 删除项目
      parameters:
      - description: Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Project'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.JsonResult'
  /project/listProjects:
    post:
      consumes:
      - application/json
      description: 查询所有项目
      parameters:
      - description: Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Project'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.JsonResult'
  /project/pageProjects:
    post:
      consumes:
      - application/json
      description: 分页查询项目
      parameters:
      - description: Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Project'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.JsonResult'
  /project/updateProject:
    put:
      consumes:
      - application/json
      description: 更新项目
      parameters:
      - description: Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Project'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.JsonResult'
  /service/deploy:
    post:
      consumes:
      - application/json
      description: 部署
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.JsonResult'
  /service/display:
    post:
      consumes:
      - application/json
      description: 展示
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.JsonResult'
swagger: "2.0"